feat: Step 02 - Add Lazy plugin manager and necessary files for its operation
This commit marks the second step in the step-by-step configuration
process for Neovim. It includes the addition of the Lazy plugin manager
and all necessary files for its correct operation in our system. The
configuration files updated in this commit are:
- lua/plugins/list.lua
- lua/plugins/lazy.lua
- lua/plugins/custom_plugins.lua
- lua/lib/icons.lua
- lua/lib/util.lua

Key changes include:
- Addition of the Lazy plugin manager.
- Configuration of Lazy plugin manager settings.
- Definition of icons for UI elements in icons.lua.
- Utility functions in util.lua for installing LSP servers only if their
corresponding executables are installed (e.g., gopls only if go is
installed).
- Custom plugins in custom_plugins.lua for installing additional
specific plugins.

As an example of the functionality, we have installed the 'mini.animate'
plugin to demonstrate the plugin management capabilities of Lazy.