Step 03: Add whichkey.lua and necessary files for its functionality
his step adds the configuration and necessary files for the
functionality of the `which-key.nvim` plugin, including the definition
of the leader key and extension with custom options. Detailed
documentation has also been added for each modified or added file.

### Added

- **plugins/tool/which-key.lua**: Configuration for the `which-key.nvim`
plugin.
- **lua/core/keys.lua**: Definition of the leader key necessary for the
functionality of `which-key.lua`, as well as many useful keybindings.
- **lua/custom_keys.lua**: Extension with more custom options for the
`which-key.lua` menu.
- **docs/keybindings_custom-keys.lua.md**: Documentation for
`custom-keys.lua`.
- **docs/keybindings_keys.lua.md**: Documentation for `keys.lua`.
- **docs/keybindings_which-key.lua.md**: Documentation for
`which-key.lua`.

### Changed

- **init.lua**: Now invokes `lua/core/keys.lua` to configure
keybindings.
- **plugins/list.lua**: Now includes the installation of
`which-key.nvim`.